User added

+++ Backend/src/com/hackathon/Location.java
@@ -16,7 +16,7 @@ public class Location {
         return longitude;
     }
 
-    public void setLongitude(int longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 
@@ -24,7 +24,7 @@ public class Location {
         return latitude;
     }
 
-    public void setLatitude(int latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
